Bump criterion to 0.3

+++ benches/benchmarks.rs
@@ -124,7 +124,7 @@ fn encode_benchmarks(byte_sizes: &[usize]) -> ParameterizedBenchmark<usize> {
     ParameterizedBenchmark::new("encode", do_encode_bench, byte_sizes.iter().cloned())
         .warm_up_time(std::time::Duration::from_millis(500))
         .measurement_time(std::time::Duration::from_secs(3))
-        .throughput(|s| Throughput::Bytes(*s as u32))
+        .throughput(|s| Throughput::Bytes(*s as u64))
         .with_function("encode_display", do_encode_bench_display)
         .with_function("encode_reuse_buf", do_encode_bench_reuse_buf)
         .with_function("encode_slice", do_encode_bench_slice)
@@ -135,7 +135,7 @@ fn decode_benchmarks(byte_sizes: &[usize]) -> ParameterizedBenchmark<usize> {
     ParameterizedBenchmark::new("decode", do_decode_bench, byte_sizes.iter().cloned())
         .warm_up_time(std::time::Duration::from_millis(500))
         .measurement_time(std::time::Duration::from_secs(3))
-        .throughput(|s| Throughput::Bytes(*s as u32))
+        .throughput(|s| Throughput::Bytes(*s as u64))
         .with_function("decode_reuse_buf", do_decode_bench_reuse_buf)
         .with_function("decode_slice", do_decode_bench_slice)
 }
